+++ packages/components/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @react-email/components
 
+## 1.0.0-canary.0
+
+### Major Changes
+
+- 9360e39: tailwind: update to using tailwindcss@v4
+
+### Patch Changes
+
+- Updated dependencies [9360e39]
+  - @react-email/tailwind@2.0.0-canary.0
+
+## 0.5.7
+
+### Patch Changes
+
+- Updated dependencies [4205615]
+- Updated dependencies [02089b1]
+  - @react-email/render@1.4.0
+  - @react-email/markdown@0.0.16
+  - @react-email/body@0.1.0
+  - @react-email/button@0.2.0
+  - @react-email/code-block@0.1.0
+  - @react-email/code-inline@0.0.5
+  - @react-email/column@0.0.13
+  - @react-email/container@0.0.15
+  - @react-email/font@0.0.9
+  - @react-email/head@0.0.12
+  - @react-email/heading@0.0.15
+  - @react-email/hr@0.0.11
+  - @react-email/html@0.0.11
+  - @react-email/img@0.0.11
+  - @react-email/link@0.0.12
+  - @react-email/preview@0.0.13
+  - @react-email/row@0.0.12
+  - @react-email/section@0.0.16
+  - @react-email/tailwind@1.2.2
+  - @react-email/text@0.1.5
+
 ## 0.5.6
 
 ### Patch Changes